% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgpss_fields.R
\name{cgpss_excluded_fields}
\alias{cgpss_excluded_fields}
\title{CGPSS Excluded Fields}
\usage{
cgpss_excluded_fields()
}
\description{
Return a list of non-question CGPSS fields.
}
\details{
In order to exclude non-question bits from the response
table, we need a list of fields that aren't questions.
}
