% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgpss_fields.R
\name{cgpss_respondent_characteristics}
\alias{cgpss_respondent_characteristics}
\title{CGPSS Respondent Characteristics}
\usage{
cgpss_respondent_characteristics()
}
\description{
Return a list of fields describing CGPSS respondents
}
