% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_groups.R
\name{compare_two_groups}
\alias{compare_two_groups}
\title{Compare Two Groups}
\usage{
compare_two_groups(df, comp_id, ...)
}
\arguments{
\item{df}{The survey data frame}

\item{comp_id}{The column of the data frame that separates the
data into two groups.}

\item{...}{The variables to group the dataset by.}
}
\value{
A data frame with the statistical comparisons.
}
\description{
Compare two groups within the data.
}
\details{
This function statistically compares two groups
and saves that information into a data frame.
}
