% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statistical_significance.R
\name{create_statistical_significance}
\alias{create_statistical_significance}
\title{Create Statistical Significance}
\usage{
create_statistical_significance(df)
}
\arguments{
\item{df}{The data frame containing `p.value`.}
}
\value{
The data frame with a `significance` column
  added.
}
\description{
Create a star indicator based on the p.value.
}
\details{
Statistical tests return p-values (in a column called
p.value - if you use the `broom` package) and
this creates the variable `significance` that
has the stars.
}
