% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_git_commit.R
\name{current_git_commit}
\alias{current_git_commit}
\title{Get the current git commit}
\usage{
current_git_commit(docker = TRUE)
}
\arguments{
\item{docker}{If `TRUE` the default, will make sure you
are in the Git repository before checking for the
latest hash. If `FALSE`, we assume you are there.}
}
\value{
The SHA-1 hash of the current Git Commit
}
\description{
Return the hash of the current git commit.
We use this for adding to the report or the database.
}
\examples{
current_git_commit()
}
