% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_sample_size.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Get Sample Size}
\usage{
get_sample_size(df)
}
\arguments{
\item{df}{The data frame containg the survey data.}
}
\value{
A data frame with the count of the two groups.
}
\description{
Get sample sizes for two groups.
}
\details{
This function finds the sample size for two
groups that are under comparison. It currently
uses `value` to separate out the two groups.
}
