% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_statistical_test.R
\name{pairwise_statistical_tests}
\alias{pairwise_statistical_tests}
\title{Pairwise Statistical Tests}
\usage{
pairwise_statistical_tests(df, comp_id)
}
\arguments{
\item{df}{The data frame containing the survey data.}

\item{comp_id}{The column which identifies the groups
in the data that you want to compare against each
other.}
}
\value{
A data frame of all the statistical comparisons.
}
\description{
Run statistical test between each combination of groups
}
\details{
Given a grouping variable (`comp_id`), go through all
possible combinations of pairs and perform statistical
tests on each one.
}
