% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_statistical_test.R
\name{pairwise_test}
\alias{pairwise_test}
\title{Pairwise Test}
\usage{
pairwise_test(df, comp_id, pair)
}
\arguments{
\item{df}{the data frame}

\item{comp_id}{the field used to break up the data into groups}

\item{pair}{a list of two groups that we want to compare}
}
\value{
a data frame containing the statistics
}
\description{
Run a Single Pairwise test
}
\details{
A helper function for `pairwise_statistical_tests`.
}
