% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respondent_table.R
\name{respondent_table}
\alias{respondent_table}
\title{Respondent Table}
\usage{
respondent_table(df, respondent_characteristic_fields, survey_id = NULL,
  respondent_id_field = NULL, survey_year_field = NULL)
}
\arguments{
\item{df}{The survey data frame. It should be labelled data.}

\item{respondent_characteristic_fields}{A list of field names
that represent characteristics of the respondent. These are
fields that you wish to include in the respondent table.}

\item{survey_id}{A string representing the name of the survey.
(e.g. "CGPSS")}

\item{respondent_id_field}{The field in the data frame that
uniquely identifies each respondent within that cycle of
the survey.}

\item{survey_year_field}{The field in the data frame that
identifies the year of the survey.}
}
\value{
A tibble containing the characteristics of each respondent
  of the survey. Each respondent will be uniquely identified by
  the ID of the survey, the year of the survey and the respondent
  identifier.
}
\description{
Create a table with one row for each respondent.
}
\details{
Extract respondent data from the survey file.
Then we make sure there is enough information to
uniquely identify each respondent.
}
\examples{
respondent_table(cgpss,
                 c('LANGUAGE', 'UNIV', 'PROGRAM_NAME', 'CIP'),
                 survey_id = 'CGPSS',
                 respondent_id_field = UNIV_RecordNum,
                 survey_year_field = SURVEY_YEAR)
}
