% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_table.R
\name{response_table}
\alias{response_table}
\title{Response Table}
\usage{
response_table(df, respondent_id_field = NULL,
  survey_year_field = NULL, excluded_fields, survey_id = NULL)
}
\arguments{
\item{df}{The survey data frame. It should have variable labels.}

\item{respondent_id_field}{The field in the data frame that
uniquely identifies each respondent within that cycle of
the survey.}

\item{survey_year_field}{the field in the data frame that
identifies the year of the survey.}

\item{excluded_fields}{A list of field names that are not
questions and shouldn't be included.}

\item{survey_id}{A string representing the name of the survey.}
}
\value{
A tibble containing all the responses of the survey,
  with the actual response, a character label and a score.
}
\description{
Create a table with one row for each response.
}
\details{
Extract the answers/responses from the survey file.
Then we create a table where there is one row for
each question per respondent -- assuming they answered
the question.
}
\examples{
response_table(cgpss,
               survey_id = 'CGPSS',
               respondent_id_field = UNIV_RecordNum,
               survey_year_field = SURVEY_YEAR,
               excluded_fields = c('LANGUAGE', 'UNIV', 'PROGRAM_NAME', 'CIP'))
}
