% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_statistical_test.R
\name{reverse_pairwise_comparison}
\alias{reverse_pairwise_comparison}
\title{Reverse Pairwise Comparisons}
\usage{
reverse_pairwise_comparison(df)
}
\arguments{
\item{df}{The dataframe to flip}
}
\value{
the flipped data frame
}
\description{
Take a Pairwise Test and Reverse it
}
\details{
This lets us do half the work of comparing
between groups.
}
