% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_effect_size.R
\name{safe_effect_size}
\alias{safe_effect_size}
\title{Safe Effect Size}
\usage{
safe_effect_size(df, outcome = score, group = value)
}
\arguments{
\item{df}{The data frame.}

\item{outcome}{The name of a numeric column that is the
outcome you are testing. (Default `score`).}

\item{group}{A column in the data frame that will separate
the data into two separate groups. (Default `value`).}
}
\description{
Safely Calculate Effect Size by Group
}
\details{
Calculate effect size on a set of data not worrying about
errors. This will return the results in a friendly
manner using the `broom` package -- or it will return
`NA`.
}
