% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_sample_size.R
\name{safe_sample_size}
\alias{safe_sample_size}
\title{Safe Sample Size}
\usage{
safe_sample_size(df)
}
\arguments{
\item{df}{The survey data frame.}
}
\value{
A tibble with the sample size.
}
\description{
Safely get sample sizes for two groups.
}
\details{
This function gets the sample size for two
groups and is robust to errors (such as one
of the groups not existing).
}
