% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_stats.R
\name{safe_stats}
\alias{safe_stats}
\title{Safe Statistics}
\usage{
safe_stats(df)
}
\arguments{
\item{df}{Survey data frame.}
}
\value{
statistics
}
\description{
Generate a set of comparison statistics safely
}
\details{
Comparison statistics ready.
}
