% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_ttest.R
\name{safe_ttest}
\alias{safe_ttest}
\title{Safe T-Test}
\usage{
safe_ttest(df)
}
\arguments{
\item{df}{The data frame.}

\item{outcome}{The name of a numeric column that is the
outcome you are testing. (Default `score`).}

\item{group}{A column in the data frame that will separate
the data into two separate groups. (Default `value`).}
}
\description{
Safely Calculate T tests by Group
}
\details{
Calculate t-tests on a set of data not worrying about
errors. This will return the results in a friendly
manner using the `broom` package -- or it will return
`NA`.
}
