% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_groups.R
\name{compare_two_groups}
\alias{compare_two_groups}
\title{Compare Two Groups}
\usage{
compare_two_groups(df, fac, fac2, ...)
}
\arguments{
\item{df}{The survey data frame}

\item{fac}{The faculty code of the first faculty you are comparing.}

\item{fac2}{The faculty code you are comparing `fac` with.}

\item{...}{The variables to group the dataset by.}
}
\value{
A data frame with the statistical comparisons.
}
\description{
Compare two groups within the data.
}
\details{
This function statistically compares two groups
and saves that information into a data frame.
}
