% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_table.R
\name{question_table}
\alias{question_table}
\title{Question Table}
\usage{
question_table(df, struct)
}
\description{
Extract the text of each question from the labelled
data loaded from the SPSS file.
}
